pyproject.toml
[tool.poetry]
name = "session-away-backend"
version = "0.1.0"
description = ""
authors = ["junhcha <quffl9612@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.100.0" }
SQLAlchemy = "^2.0.17"
alembic = "^1.11.1"
pytest = "^7.4.0"
pydantic = "^2.0.0"
asyncpg = "^0.28.0"
fastapi-users = { extras = ["oauth2", "sqlalchemy"], version = "^12.1.0" }
loguru = "^0.7.0"
psycopg2 = "^2.9.6"
httpx-oauth = "^0.13.0"
asgi-lifespan = "^2.1.0"
pytest-asyncio = "^0.21.1"
sqlalchemy-utils = "^0.41.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.4.1"

[tool.mypy]
max-line-length = 88

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = ['--asyncio-mode=auto']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
